requirements.txt
# HTTP client
httpx>=0.27.0,<0.28.0

# HTML parsing
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.1.0,<6.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Logging
loguru>=0.7.0,<0.8.0

# Retry logic
tenacity>=8.2.0,<9.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Yahoo Finance
yfinance>=0.2.36,<0.3.0

# Data processing
pandas>=2.1.0,<3.0.0

# Google Gemini
google-genai>=0.3.0,<1.0.0

# API framework
fastapi>=0.109.0,<0.120.0
uvicorn[standard]>=0.27.0,<0.30.0

# Lambda adapter
mangum>=0.17.0,<0.18.0

# Testing (dev dependencies)
pytest>=8.0.0,<9.0.0
pytest-cov>=4.1.0,<5.0.0
ruff>=0.2.0,<0.3.0
